+++ _tools/check_docs.ts
@@ -48,6 +48,7 @@ const ENTRY_POINTS = [
   "../encoding/unstable_base64_stream.ts",
   "../encoding/unstable_base32hex_stream.ts",
   "../encoding/unstable_base32_stream.ts",
+  "../encoding/unstable_base32hex_stream.ts",
   "../expect/mod.ts",
   "../fmt/bytes.ts",
   "../fmt/colors.ts",
