+++ encoding/deno.json
@@ -6,7 +6,7 @@
     "./ascii85": "./ascii85.ts",
     "./base32": "./base32.ts",
     "./unstable-base32-stream": "./unstable_base32_stream.ts",
-    "./base32hex": "./base32hex.ts",
+    "./unstable-base32hex": "./unstable_base32hex.ts",
     "./unstable-base32hex-stream": "./unstable_base32hex_stream.ts",
     "./base58": "./base58.ts",
     "./base64": "./base64.ts",
