+++ encoding/unstable_base32hex_test.ts
@@ -2,7 +2,7 @@
 // Copyright (c) 2016-2017 Linus Unnebäck. MIT license.
 // Copyright 2018-2024 the Deno authors. All rights reserved. MIT license.
 import { assertEquals, assertExists, assertThrows } from "@std/assert";
-import { decodeBase32Hex, encodeBase32Hex } from "./base32hex.ts";
+import { decodeBase32Hex, encodeBase32Hex } from "./unstable_base32hex.ts";
 
 // Test vectors from https://www.rfc-editor.org/rfc/rfc4648.html#section-10
 const testCases = [
